*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@149 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tas/web/Query.java
@@ -28,7 +28,7 @@ public class Query
 			return false;
 		
 		QueryCondition queryCondition = null;
-		switch (attribute.getType())
+		switch (attribute.getType().intValue())
 		{
 			case AbstractAttribute.TYPE_STRING:
 				queryCondition = new QueryConditionText(attribute);
